tree add0b1ccffa1a0eb43c17781a385caad1b147fa9 author Brian Drennan <brian.e.drennan@outlook.com> 1669870882 -0500 committer Brian Drennan <brian.e.drennan@outlook.com> 1669870882 -0500
Initial stab at Protohackers 0

This is my first forray into socket programming based on the problem
sets from the Protohackers website. This first change provides an
implementation of RFC 862 (TCP echo) based on the problem description
located here:

- https://protohackers.com/problem/0

The code is a bit messy. I haven't looked at socket APIs in years. I
also had a hard time getting the Windows Telnet client to work well
for testing the server, so I created a custom client. The client
supports reading data from the standard input stream to send to the
server, but will fall back to prompting the user for input if the
standard input stream has not been redirected. This provides a very
Unix-y experience for transmitting data to the server. The server
socket listener is on port 7. There is not currently a means for
overriding that. It's also not terribly robust or resilient on either
side. That may come in the future if I decide to continue down this
path.