Use System.IO.Pipelines in SmokeTest
The previous version of SmokeTest did all the buffer management stuff
internally. This was fine to get an example up and running, but isn't
great for a proper echo service. This version now uses the pipelines
library provided by System.IO.Pipelines, which handles buffer
management on behalf of the user. This is a much more robust
implementation because it allows the server to receive a lot more
data without any up-front planning about the size of that data, while
also maintaining a reasonably well-readable codebase. It also stops
logging the entirety of the messages that were delivered and instead
just logs the size of the data transferred (in both directions).

This also highlighted a bug in the client where the client expected to
receive the entirety of the message it transferred in a single response
message. The client is now updated to keep track of the total length of
the message it sent, and continue calling `Receive()` on its' socket
until that many bytes have been sent back to it.

This was all tested using the source code of the roslyn repo in a local
test via the following pwsh commands:

```PowerShell
Get-Item $rosly | Where-Object PSIsContainer -eq $false |
    ForEach-Object {
        "Sending file content {0} (Length = {1}" -f $_.FullName, $_.Length
        Get-Content $_.FullName | .\TcpClient.exe 127.0.0.1 7
    }
```