Fix buffer sizing in SmokeTest
The previous version of this code requested buffers with a size hint
that was completely arbitrary (and unnecessary). This change removes
the use of size hints when requesting a buffer and delegates that
entire task to the framework, which has smart heuristics already built
in for determining the appropriate size of a buffer.